more refactoring: regroup streamtools package files, removed blocks which haven't been converted
