+++ streamtools/block_functions.go
