+++ streamtools/blocks_basics.go
